requirements.txt
# EDS Early Detection System Requirements
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0
flask==2.3.3
flask-cors==4.0.0
dash==2.14.1
plotly==5.16.1
pandas==2.1.1
numpy==1.24.3
geopandas==0.14.0
rasterio==1.3.8
shapely==2.0.1
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.2
redis==4.6.0
gunicorn==21.2.0
schedule==1.2.0
pydantic==2.3.0
fastapi==0.103.1
uvicorn==0.23.2
boto3==1.28.57